Coffee Shop Command Line Interface Program
OKCoders program to finish off our third module! Allows user to type what they want from a list, hit "checkout" to see how many items were ordered and the total price, and buttons to suggest food/drink based on the weather for a specified zip code on either the current day, or a future date. This isn't done *quite* yet, but I figured I could upload it as is and then update it once I've done a little more work to make it *pretty*